LPS-132025 Moving NumericDDMFormFieldUtil to api module and refactor to not reuse DecimalFormat instances, avoiding concurrency issues
